requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
Django==2.1.7
django-extensions==2.1.6
django-taggit==1.3.0
graphviz==0.10.1
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
Pillow==5.4.1
pycparser==2.19
pydot==1.4.1
pyparsing==2.4.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
werkzeug==0.15.2
wincertstore==0.2
